+++ CHANGELOG.md
@@ -1,8 +1,19 @@
-# tartiflette-plugin-isodate_tartiflette_plugin
+# tartiflette-plugin-isodatenow
 
 ## Released Versions
 
-### [x.y.Z]
+### [0.2.0]
+- Renamed `isoDate` to `isoDateNow`
+
+#### Fixed
+
+#### Changed
+- Renamed `isoDate` to `isoDateNow`
+
+#### Added
+
+### [0.1.4]
+- Initial release
 
 #### Fixed